Add better logging and clubrole endpoint

+++ server/controllers/clubController.js
@@ -1,17 +1,20 @@
 const Club = require("../models/clubModel");
+const clubRole = require("./clubroleController");
 
 exports.createClub = async(req, res) => {
     try {
-
+        console.log(`${req.sessionID} - ${req.session.email} is requesting to create club. Changes: ${JSON.stringify(req.body)}`);
         const { name, description, email } = req.body;
         //const executives = [];
+
+        if (!req.session.isLoggedIn) {
+            throw new Error('Unauthorized: Must sign in to create a club');
+        }
 
         // Validate email format
         const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
         if (!emailRegex.test(email)) {
-            return res.status(400).json({ 
-                message: "Invalid email format",
-            });;
+            throw new Error('Bad Request: Invalid email format');
         }
 
         const newClub = await Club.create({
@@ -21,20 +24,40 @@ exports.createClub = async(req, res) => {
         });
 
         res.status(200).json({ message: 'Club created successfully' });
-        
+        console.log(`${req.sessionID} - Request Success: ${req.method}  ${req.originalUrl}`);
+
     } catch (err) {
-        console.error('Club creation failed:', err);
-        res.status(500).json({ message: 'An error occurred while processing your request' });
+        if (err.message.includes('Unauthorized')) {
+            res.status(403).json({
+                status: "fail",
+                message: err.message,
+                description: `Unauthorized: ${req.session.email} is not an account`
+            });
+        } else if (err.message.includes('Bad Request')) {
+            res.status(400).json({
+                status: "fail",
+                message: err.message,
+                description: `Bad Request: Failed to create club`
+            });
+        } else {
+            res.status(500).json({
+                status: "fail",
+                message: err.message,
+                description: `Bad Request: Server Error`
+            });
+            console.log(`${req.sessionID} - Server Error: ${err}`)
+        }
+        console.log(`${req.sessionID} - Request Failed: ${err.message}`);
     }
 };
 
-// Example: get a club from the DB based on the name passed in the request json
 exports.getClub = async(req, res) => {
     try {
+        console.log(`${req.sessionID} - Request Get on ${ req.params.name}`);
         const club = await Club.findOne({ name: req.params.name });
 
-        if (club == null) {
-            throw err;
+        if (!club) {
+            throw new Error('Not Found: Fail to edit club as DNE');
         }
         res.status(200).json({
             status: "success",
@@ -43,25 +66,50 @@ exports.getClub = async(req, res) => {
                 club: club,
             },
         });
-       
+        console.log(`${req.sessionID} - Request Success: ${req.method}  ${req.originalUrl}`);
+
     } catch (err) {
-        res.status(404).json({
-            status: "fail",
-            message: err.message,
-            description: "not found club",
-        });
+        if (err.message.includes('Not Found')) {
+            res.status(404).json({
+                status: "fail",
+                message: err.message,
+                description: `Not Found: Fail to edit club as ${req.params.name} DNE`,
+            });
+        } else {
+            res.status(500).json({
+                status: "fail",
+                message: err.message,
+                description: `Bad Request: Server Error`,
+            });
+            console.log(`${req.sessionID} - Server Error: ${err}`)
+        }
+        console.log(`${req.sessionID} - Request Failed: ${err.message}`);
     }
 };
 
 exports.editClub = async(req, res) => {
     try {
-        console.log(req.body);
-        const club = await Club.updateOne({ name: req.params.name },req.body);
-        
-        if (club == null) {
-            throw err;
+        console.log(`${req.sessionID} - ${req.session.email} is requesting to edit club ${ req.params.name}. Changes: ${JSON.stringify(req.body)}`);
+        // Checking if club exists first as we need a valid club to get possible role
+        const club = await Club.findOne({ name: req.params.name });  
+        if (!club) {
+            throw new Error('Not Found: Fail to edit club as DNE');
+        }
+
+        if (!req.session.isLoggedIn) {
+            throw new Error('Unauthorized: Must sign in to edit a club');
         }
         
+        const isAdmin = await clubRole.isClubAdminMiddleware(req.session.email, req.params.name);
+        if (!isAdmin) {
+            throw new Error('Unauthorized: Only admins can modify the club.');
+        }
+        
+        const updateStatus = await Club.updateOne({ name: req.params.name },req.body);
+        if (!updateStatus.acknowledged) {
+            throw err;
+        }
+
         res.status(201).json({
             status: "success",
             message: "club modified",
@@ -69,35 +117,86 @@ exports.editClub = async(req, res) => {
                 club: club,
             },
         });
+        console.log(`${req.sessionID} - Request Success: ${req.method}  ${req.originalUrl}`);
+
     } catch (err) {
-        res.status(400).json({
-            status: "fail",
-            message: err.message,
-            description: "Fail to create new club",
-        });
+        if (err.message.includes('Unauthorized')) {
+            res.status(403).json({
+                status: "fail",
+                message: err.message,
+                description: `Unauthorized: ${req.session.email} is not and admin of club ${req.params.name}`,
+            });
+        } else if (err.message.includes('Not Found')) {
+            res.status(404).json({
+                status: "fail",
+                message: err.message,
+                description: `Not Found: Fail to edit club as ${req.params.name} DNE`,
+            });
+        } else {
+            res.status(500).json({
+                status: "fail",
+                message: err.message,
+                description: `Bad Request: Server Error`,
+            });
+            console.log(`${req.sessionID} - Server Error: ${err}`)
+        }
+        console.log(`${req.sessionID} - Request Failed: ${err.message}`);
     }
 };
 
 exports.deleteClub = async(req, res) => {
     try {
-        const club = await Club.deleteMany({ name: req.params.name });
+        console.log(`${req.sessionID} - ${req.session.email} is requesting to delete club ${ req.params.name}`);
+        // Checking if club exists first as we need a valid club to get possible role
+        const club = await Club.findOne({ name: req.params.name }); 
+         if (!club) {
+            throw new Error('Not Found: Fail to edit club as DNE');
+        }
 
-        if (club == null) {
+        if (!req.session.isLoggedIn) {
+            throw new Error('Unauthorized: Must sign in to delete a club');
+        }
+
+        const isAdmin = await clubRole.isClubAdminMiddleware(req.session.email, req.params.name);
+        if (!isAdmin) {
+            throw new Error('Unauthorized: Only admins can delete the club.');
+        }
+
+        const deleteStatus = await Club.deleteMany({ name: req.params.name });
+        if (!updateStatus.acknowledged) {
             throw err;
         }
+
         res.status(200).json({
             status: "success",
-            message: "Found club",
+            message: "club deleted",
             data: {
                 club: club,
             },
         });
-       
+        console.log(`${req.sessionID} - Request Success: ${req.method}  ${req.originalUrl}`);
+
     } catch (err) {
-        res.status(404).json({
-            status: "fail",
-            message: err.message,
-            description: "not found club",
-        });
+        if (err.message.includes('Unauthorized')) {
+            res.status(403).json({
+                status: "fail",
+                message: err.message,
+                description: `Unauthorized: ${req.session.email} is not and admin of club ${req.params.name}`,
+            });
+        } else if (err.message.includes('Not Found')) {
+            res.status(404).json({
+                status: "fail",
+                message: err.message,
+                description: `Not Found: Fail to delete new club as ${req.params.name} DNE`,
+            });
+        } else {
+            res.status(500).json({
+                status: "fail",
+                message: err.message,
+                description: `Bad Request: Server Error`,
+            });
+            console.log(`${req.sessionID} - Server Error: ${err}`)
+        }
+        console.log(`${req.sessionID} - Request Failed: ${err.message}`);
     }
 };
